software/requirements.txt
pytorch-lightning
transformers
toml
timm
pytorch-nlp
datasets
onnx
black
GitPython
colorlog
cocotb[bus]
pytest
ipython
ipdb
sentencepiece
einops
deepspeed
pybind11
tabulate
tensorboard